Prepare file structure to support `next-intl`
