+++ next.config.mjs
@@ -1,4 +1,8 @@
+import createNextIntlPlugin from 'next-intl/plugin'
+
+const withNextIntl = createNextIntlPlugin('./src/libs/i18n.ts')
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {}
 
-export default nextConfig
+export default withNextIntl(nextConfig)
