+++ src/app/layout.tsx
@@ -1,20 +1,9 @@
-import './globals.css'
-
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: 'Adam Ajsaen',
-  description: 'My personal website',
-}
+import '@/styles/globals.css'
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode
-}>) {
-  return (
-    <html lang="en" suppressHydrationWarning={true}>
-      <body>{children}</body>
-    </html>
-  )
+}) {
+  return children
 }
