+++ src/messages/en.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "Hello world!"
+  }
+}
