+++ src/messages/th.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "สวัสดีครับ!"
+  }
+}
